Refactor de bodyParser
